// src/constantes/types.ts
export enum PokemonTypes {
    bug = 'bug',
    fire = 'fire',
    dark = 'dark',
    dragon = 'dragon',
    fairy = 'fairy',
    electric = 'electric',
    fighting = 'fighting',
    flying = 'flying',
    ghost = 'ghost',
    grass = 'grass',
    ground = 'ground',
    ice = 'ice',
    normal = 'normal',
    poison = 'poison',
    psychic = 'psychic',
    rock = 'rock',
    steel = 'steel',
    water = 'water',
}